requirements.txt
Flask==2.1.1
python-dotenv==0.19.2
python-geoip-geolite2==2015.0303
maxminddb==2.2.0
geoip2==4.2.0
pillow==9.5.0
gunicorn==20.1.0